Extract local-subtree-reload region into sections/local-subtree-reload.js
Fourth region extraction of P-09. 35 LOC of pure helpers — _supportsFragmentFetch, _captureDetailsOpenState, _restoreDetailsOpenState. No cross-region state so the module needs no imports and has no side effects.
